requirements.txt
cx-Freeze==6.5.2
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
mysqlclient==2.0.1
PyMySQL==1.0.2
SQLAlchemy==1.3.18
Werkzeug==1.0.1
PyJWT==1.7.1
simplejson==3.16.0